+++ static/js/attraction.js
