+++ templates/attraction.html
@@ -4,7 +4,7 @@
         <meta name="viewport" content="width=device-width, initial-scale=1.0" />
         <meta charset="utf-8"/>   
         <title>Taipei Attraction</title>
-        <link rel="stylesheet" type="text/css" href="/static/attraction.css" />
+        <link rel="stylesheet" type="text/css" href="/static/css/attraction.css" />
     </head>
     <body>
         <div class="content">
@@ -66,6 +66,6 @@
             </div>
         </div> 
         <div class="footer"></div>
-        <script src="/static/attraction.js"></script>
+        <script src="/static/js/attraction.js"></script>
     </body>
 </html>
