+++ templates/index.html
@@ -4,7 +4,7 @@
         <meta name="viewport" content="width=device-width, initial-scale=1.0" />
         <meta charset="utf-8"/>   
         <title>Taipei Day Trip</title>
-        <link rel="stylesheet" type="text/css" href="/static/index.css" />
+        <link rel="stylesheet" type="text/css" href="/static/css/index.css" />
     </head>
     <body>
         <div class="content">
@@ -34,6 +34,6 @@
             </div>
         </div>    
         <div class="footer"></div>
-        <script src="/static/index.js"></script>
+        <script src="/static/js/index.js"></script>
     </body>
 </html>
